Refactor folders structure
