+++ idea/frontend/src/features/voucher/consts.ts
