+++ idea/frontend/src/features/voucher/hooks/use-duration-schema.ts
