+++ idea/frontend/src/features/voucher/hooks/use-voucher-type.ts
