+++ idea/frontend/src/features/voucher/ui/duration-form/index.ts
@@ -1,0 +1,3 @@
+import { DurationForm } from './duration-form';
+
+export { DurationForm };
