+++ idea/frontend/src/features/voucher/ui/programs-form/index.ts
@@ -1,0 +1,3 @@
+import { ProgramsForm } from './programs-form';
+
+export { ProgramsForm };
