+++ idea/frontend/src/features/voucher/ui/programs-form/programs-form.module.scss
@@ -1,0 +1,39 @@
+.addProgram {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+
+  .input {
+    display: flex;
+    align-items: flex-end;
+    gap: 8px;
+
+    &.error {
+      align-items: center;
+    }
+  }
+
+  ul {
+    display: grid;
+    grid-template-columns: repeat(5, minmax(0, 1fr));
+    gap: 8px;
+  }
+
+  .programId {
+    padding: 2px 8px;
+
+    display: flex;
+    align-items: center;
+
+    font-size: 12px;
+    color: rgba(#fff, 0.8);
+
+    background: rgba(#fff, 0.1);
+    border-radius: 15px;
+
+    span {
+      text-overflow: ellipsis;
+      overflow: hidden;
+    }
+  }
+}
