+++ idea/frontend/src/features/voucher/utils.ts
